fix: postpone loading cvxpy.
Loading cvxpy the first time can be quite slow when there is no byte
cache yet; load cvxpy only when we actually need the solver or when
enumerating supported backends, so a front-end can at least load that
little bit faster.
